Hoàn chỉnh phiếu nhập (cái tổng khá sai sai) ae test thử, nhân viên run bị đỏ chí, sản phẩm thêm đỏ chí, sửa thì méo đc

+++ src/BUS/ChiTietPhieuNhapService.java
@@ -70,7 +70,7 @@ public class ChiTietPhieuNhapService {
         for (int i = result.size() - 1; i >= 0; i--) {
             ChiTietPhieuNhap ct = result.get(i);
             int sl = ct.getsLuong();
-            float tt = ct.getDonGia() * sl;
+            double tt = ct.getDonGia()* sl;
 
             Boolean soLuongKhongThoa = (soLuong1 != -1 && sl < soLuong1) || (soLuong2 != -1 && sl > soLuong2);
             Boolean donGiaKhongThoa = (thanhTien1 != -1 && tt < thanhTien1) || (thanhTien2 != -1 && tt > thanhTien2);
